Improved arrays a lot.
Now it's possible to write code like:
> tab_2d[index] = tab_1d;
and create Java-like arrays:
> var<int[][]> tab = new int[10][];

+++ ssvm/gc/garbagecollector.pas
@@ -270,6 +270,6 @@ Begin
  Result := 0;
 
  For List in ObjectLists Do
-  Result += List.Count;
+  Result += uint32(List.Count);
 End;
 End.
